Fix diffCourseSets to properly handle course changes
Before, diffCourseSets failed to sort through additions of new courses,
dropped courses, and switching courses (to different periods). These
changes fix all of that.

The original method neglected the cases where a course is added to a
period that wasn't in the `a` *Gradebook and also failed to detect new
courses added to the Gradebook.