Fix old assignments being marked as removed
New assignments are prepended to the []*Assignments slice, so old
assignments have their indexes increased by n, where n is the number of
new assignments. Originally, we only iterated over the old
[]*Assignments slice (against which we were diffing the new one), which
means that only k - n assignments were looked at, where k is the number
of assignments in the new []*Assignments slice.

The result of this is that we leave n assignments in the new
[]*Assignments slice untouched and not looked at. The result of this is
that old assignments (that had their indexes shifted) are marked as
removed since they are never seen.